+++ src/nativeMain/kotlin/com/ditchoom/buffer/BufferFactory.kt
@@ -4,12 +4,12 @@ package com.ditchoom.buffer
 
 
 actual fun PlatformBuffer.Companion.allocate(
-    size: UInt,
+    size: Int,
     byteOrder: ByteOrder
-): PlatformBuffer = NativeBuffer(ByteArray(size.toInt()), byteOrder = byteOrder)
+): PlatformBuffer = NativeBuffer(ByteArray(size), byteOrder = byteOrder)
 
 actual fun PlatformBuffer.Companion.wrap(array: ByteArray, byteOrder: ByteOrder): PlatformBuffer =
     NativeBuffer(array, byteOrder = byteOrder)
 
 actual fun String.toBuffer(): PlatformBuffer = NativeBuffer(this.encodeToByteArray(), byteOrder = ByteOrder.BIG_ENDIAN)
-actual fun String.utf8Length(): UInt = encodeToByteArray().size.toUInt()
+actual fun String.utf8Length(): Int = encodeToByteArray().size
